Restructured project to be adapterbased
